remove send result back to client

+++ src/computation/Actions.java
@@ -18,11 +18,11 @@ public class Actions {
     }
     
     // Java program for Fibonacci number using Loop.
-    public static int fibonacciLoop(int number){
+    public static long fibonacciLoop(int number){
         if(number == 1 || number == 2){
             return 1;
         }
-        int fibo1=1, fibo2=1, fibonacci=1;
+        long fibo1=1, fibo2=1, fibonacci=1;
         for(int i= 3; i<= number; i++){
             fibonacci = fibo1 + fibo2; //Fibonacci number is sum of previous two Fibonacci number
             fibo1 = fibo2;
